+++ src/main/java/fr/foodfast/tp1/Application.java
@@ -1,5 +1,10 @@
 package fr.foodfast.tp1;
 
+import fr.foodfast.tp1.delivery.DeliveryPlatform;
+import fr.foodfast.tp1.delivery.DeliveryProcessingException;
+import fr.foodfast.tp1.order.Order;
+import fr.foodfast.tp1.order.OrderPreparationException;
+
 public class Application {
 
     public static void main(String[] args) {
