+++ src/main/java/fr/foodfast/tp1/Subscriber.java
@@ -1,5 +1,7 @@
 package fr.foodfast.tp1;
 
+import fr.foodfast.tp1.event.Event;
+
 public interface Subscriber {
     void handleEvent(Event event);
 }
