+++ src/main/java/fr/foodfast/tp1/delivery/DeliveryPlatform.java
@@ -1,4 +1,9 @@
-package fr.foodfast.tp1;
+package fr.foodfast.tp1.delivery;
+import fr.foodfast.tp1.*;
+import fr.foodfast.tp1.event.Event;
+import fr.foodfast.tp1.order.Order;
+import fr.foodfast.tp1.order.OrderEvent;
+
 import java.util.*;
 
 public class DeliveryPlatform implements Subscriber {
