+++ src/main/java/fr/foodfast/tp1/delivery/DeliveryProcessingException.java
@@ -1,4 +1,4 @@
-package fr.foodfast.tp1;
+package fr.foodfast.tp1.delivery;
 
 public class DeliveryProcessingException extends Exception {
     public DeliveryProcessingException(String message) {
