+++ src/main/java/fr/foodfast/tp1/event/Event.java
@@ -1,4 +1,4 @@
-package fr.foodfast.tp1;
+package fr.foodfast.tp1.event;
 
 public interface Event {
     EventType getEventType();
