+++ src/main/java/fr/foodfast/tp1/order/OrderPreparationException.java
@@ -1,4 +1,4 @@
-package fr.foodfast.tp1;
+package fr.foodfast.tp1.order;
 
 public class OrderPreparationException extends Exception {
     public OrderPreparationException(String message) {
