+++ src/test/java/fr/foodfast/tp1/OrderTest.java
@@ -1,7 +1,8 @@
 package fr.foodfast.tp1;
 
-import org.junit.jupiter.api.Test;
-import static org.junit.jupiter.api.Assertions.*;
+import fr.foodfast.tp1.delivery.DeliveryPlatform;
+import fr.foodfast.tp1.order.Order;
+import fr.foodfast.tp1.order.OrderPreparationException;
 import org.junit.jupiter.api.Test;
 import static org.junit.jupiter.api.Assertions.*;
