+++ src/test/java/fr/foodfast/tp1/TestEventBusWithMultipleEventTypes.java
@@ -1,5 +1,9 @@
 package fr.foodfast.tp1;
 
+import fr.foodfast.tp1.delivery.DeliveryEvent;
+import fr.foodfast.tp1.event.EventBus;
+import fr.foodfast.tp1.order.Order;
+import fr.foodfast.tp1.order.OrderEvent;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
